+++ requirements.txt
@@ -11,5 +11,6 @@ pytz==2019.3
 rauth==0.7.3
 regex==2020.1.8
 requests==2.22.0
+scipy==1.6.1
 six==1.14.0
 urllib3==1.25.8
